Updated Client model and all corresponding files to include User constraint. Also, included updated database migration

+++ controllers/clientController.js
@@ -1,20 +1,23 @@
 const { Client } = require('../models/index');
 const { v4: uuid } = require('uuid');
 
-const getClients = async () => {
-    const clients = await Client.findAll();
+const getClients = async (userId) => {
+    const clients = await Client.findAll({ where: { user_id: userId } });
     return clients;
 };
 
-const getClientById = async (clientId) => {
+const getClientById = async (userId, clientId) => {
+
     const client = await Client.findByPk(clientId);
-    if (!client) return null;
+    if (!client || client.user_id !== userId) return null;
     return client;
 };
 
-const addClient = async (data) => {
+const addClient = async (userId, data) => {
+    if (!userId) throw new Error('User ID is required to create a client');
     const newClient = await Client.create({
         client_id: uuid(),
+        user_id: userId,
         email: data.email,
         first_name: data.first_name,
         last_name: data.last_name,
@@ -23,7 +26,8 @@ const addClient = async (data) => {
     return newClient;
 };
 
-const updateClient = async (clientId, data) => {
+const updateClient = async (userId, clientId, data) => {
+    if (!userId) throw new Error('User ID is required to update a client');
     const updatedClient = await Client.update(
         {
             email: data.email,
@@ -33,17 +37,20 @@ const updateClient = async (clientId, data) => {
         },
         {
             where: {
-                client_id: clientId
+                client_id: clientId,
+                user_id: userId
             }
         }
     );
     return updatedClient;
 };
 
-const deleteClient = async (clientId) => {
+const deleteClient = async (userId, clientId) => {
+    if (!userId) throw new Error('User ID is required to delete a client');
     const deletedClient = await Client.destroy({
         where: {
-            client_id: clientId
+            client_id: clientId,
+            user_id: userId
         }
     });
     return deletedClient;
